requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.17.0
appdirs==1.4.4
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
black==21.4b2
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dateparser==1.0.0
discord.py==2.0.0a3089+g1bf7aadf
flake8==3.9.1
gino==1.0.1
idna==3.1
importlib-metadata==4.0.1
kiwisolver==1.3.1
matplotlib==3.4.1
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.20.2
pathspec==0.8.1
Pillow==8.2.0
py-cpuinfo==8.0.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.15.0
SQLAlchemy==1.3.24
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
yarl==1.6.3
zipp==3.4.1
